+++ chroot.c
@@ -20,23 +20,10 @@
 #include "iamroot.h"
 
 extern char *next_getcwd(char *, size_t);
-extern int next_stat(const char *, struct stat *);
-extern int next_fstat(int, struct stat *);
-extern int next_lstat(const char *, struct stat *);
 extern int next_fstatat(int, const char *, struct stat *, int);
-extern int next___fstat(int, struct stat *);
-extern int next___xstat(int, const char *, struct stat *);
-extern int next___fxstat(int, int, struct stat *);
-extern int next___lxstat(int, const char *, struct stat *);
 extern int next___fxstatat(int, int, const char *, struct stat *, int);
 #ifdef __GLIBC__
-extern int next_stat64(const char *, struct stat64 *);
-extern int next_fstat64(int, struct stat64 *);
-extern int next_lstat64(const char *, struct stat64 *);
 extern int next_fstatat64(int, const char *, struct stat64 *, int);
-extern int next___xstat64(int, const char *, struct stat64 *);
-extern int next___fxstat64(int, int, struct stat64 *);
-extern int next___lxstat64(int, const char *, struct stat64 *);
 extern int next___fxstatat64(int, int, const char *, struct stat64 *, int);
 extern int next_statx(int, const char *, int, unsigned int, struct statx *);
 #endif
@@ -81,7 +68,7 @@ int fissymlink(int fd)
 	struct stat statbuf;
 	int ret;
 
-	ret = next_fstat(fd, &statbuf);
+	ret = next_fstatat(fd, "", &statbuf, AT_EMPTY_PATH);
 	if (ret == -1)
 		return ret;
 
@@ -94,7 +81,7 @@ int issymlink(const char *path)
 	struct stat statbuf;
 	int ret;
 
-	ret = next_lstat(path, &statbuf);
+	ret = next_fstatat(AT_FDCWD, path, &statbuf, AT_SYMLINK_NOFOLLOW);
 	if (ret == -1)
 		return ret;
 
@@ -120,7 +107,7 @@ int isdirectory(const char *path)
 	struct stat statbuf;
 	int ret;
 
-	ret = next_lstat(path, &statbuf);
+	ret = next_fstatat(AT_FDCWD, path, &statbuf, AT_SYMLINK_NOFOLLOW);
 	if (ret == -1)
 		return ret;
 
@@ -133,7 +120,7 @@ int fisdirectory(int fd)
 	struct stat statbuf;
 	int ret;
 
-	ret = next_fstat(fd, &statbuf);
+	ret = next_fstatat(fd, "", &statbuf, AT_EMPTY_PATH);
 	if (ret == -1)
 		return ret;
 
@@ -159,7 +146,7 @@ int fisfile(int fd)
 	struct stat statbuf;
 	int ret;
 
-	ret = next_fstat(fd, &statbuf);
+	ret = next_fstatat(fd, "", &statbuf, AT_EMPTY_PATH);
 	if (ret == -1)
 		return ret;
 
@@ -172,7 +159,7 @@ int isfile(const char *path)
 	struct stat statbuf;
 	int ret;
 
-	ret = next_lstat(path, &statbuf);
+	ret = next_fstatat(AT_FDCWD, path, &statbuf, AT_SYMLINK_NOFOLLOW);
 	if (ret == -1)
 		return ret;
 
@@ -239,75 +226,6 @@ int inchroot()
 }
 
 __attribute__((visibility("hidden")))
-int rootstat(const char *path, struct stat *buf)
-{
-	uid_t uid;
-	gid_t gid;
-	int ret;
-
-	ret = next_stat(path, buf);
-	if (ret == -1)
-		goto exit;
-
-	uid = next_geteuid();
-	if (buf->st_uid == uid)
-		buf->st_uid = geteuid();
-
-	gid = getegid();
-	if (buf->st_gid == gid)
-		buf->st_gid = 0;
-
-exit:
-	return ret;
-}
-
-__attribute__((visibility("hidden")))
-int rootfstat(int fd, struct stat *buf)
-{
-	uid_t uid;
-	gid_t gid;
-	int ret;
-
-	ret = next_fstat(fd, buf);
-	if (ret == -1)
-		goto exit;
-
-	uid = next_geteuid();
-	if (buf->st_uid == uid)
-		buf->st_uid = geteuid();
-
-	gid = getegid();
-	if (buf->st_gid == gid)
-		buf->st_gid = 0;
-
-exit:
-	return ret;
-}
-
-__attribute__((visibility("hidden")))
-int rootlstat(const char *path, struct stat *buf)
-{
-	uid_t uid;
-	gid_t gid;
-	int ret;
-
-	ret = next_lstat(path, buf);
-	if (ret == -1)
-		goto exit;
-
-	uid = next_geteuid();
-	if (buf->st_uid == uid)
-		buf->st_uid = geteuid();
-
-	gid = getegid();
-	if (buf->st_gid == gid)
-		buf->st_gid = 0;
-
-exit:
-	return ret;
-}
-
-__attribute__((visibility("hidden")))
 int rootfstatat(int fd, const char *path, struct stat *buf, int flags)
 {
 	uid_t uid;
@@ -315,75 +233,6 @@ int rootfstatat(int fd, const char *path, struct stat *buf, int flags)
 	int ret;
 
 	ret = next_fstatat(fd, path, buf, flags);
-	if (ret == -1)
-		goto exit;
-
-	uid = next_geteuid();
-	if (buf->st_uid == uid)
-		buf->st_uid = geteuid();
-
-	gid = getegid();
-	if (buf->st_gid == gid)
-		buf->st_gid = 0;
-
-exit:
-	return ret;
-}
-
-__attribute__((visibility("hidden")))
-int __rootxstat(int ver, const char *path, struct stat *buf)
-{
-	uid_t uid;
-	gid_t gid;
-	int ret;
-
-	ret = next___xstat(ver, path, buf);
-	if (ret == -1)
-		goto exit;
-
-	uid = next_geteuid();
-	if (buf->st_uid == uid)
-		buf->st_uid = geteuid();
-
-	gid = getegid();
-	if (buf->st_gid == gid)
-		buf->st_gid = 0;
-
-exit:
-	return ret;
-}
-
-__attribute__((visibility("hidden")))
-int __rootfxstat(int ver, int fd, struct stat *buf)
-{
-	uid_t uid;
-	gid_t gid;
-	int ret;
-
-	ret = next___fxstat(ver, fd, buf);
-	if (ret == -1)
-		goto exit;
-
-	uid = next_geteuid();
-	if (buf->st_uid == uid)
-		buf->st_uid = geteuid();
-
-	gid = getegid();
-	if (buf->st_gid == gid)
-		buf->st_gid = 0;
-
-exit:
-	return ret;
-}
-
-__attribute__((visibility("hidden")))
-int __rootlxstat(int ver, const char *path, struct stat *buf)
-{
-	uid_t uid;
-	gid_t gid;
-	int ret;
-
-	ret = next___lxstat(ver, path, buf);
 	if (ret == -1)
 		goto exit;
 
@@ -425,75 +274,6 @@ exit:
 
 #ifdef __GLIBC__
 __attribute__((visibility("hidden")))
-int rootstat64(const char *path, struct stat64 *buf)
-{
-	uid_t uid;
-	gid_t gid;
-	int ret;
-
-	ret = next_stat64(path, buf);
-	if (ret == -1)
-		goto exit;
-
-	uid = next_geteuid();
-	if (buf->st_uid == uid)
-		buf->st_uid = geteuid();
-
-	gid = getegid();
-	if (buf->st_gid == gid)
-		buf->st_gid = 0;
-
-exit:
-	return ret;
-}
-
-__attribute__((visibility("hidden")))
-int rootfstat64(int fd, struct stat64 *buf)
-{
-	uid_t uid;
-	gid_t gid;
-	int ret;
-
-	ret = next_fstat64(fd, buf);
-	if (ret == -1)
-		goto exit;
-
-	uid = next_geteuid();
-	if (buf->st_uid == uid)
-		buf->st_uid = geteuid();
-
-	gid = getegid();
-	if (buf->st_gid == gid)
-		buf->st_gid = 0;
-
-exit:
-	return ret;
-}
-
-__attribute__((visibility("hidden")))
-int rootlstat64(const char *path, struct stat64 *buf)
-{
-	uid_t uid;
-	gid_t gid;
-	int ret;
-
-	ret = next_lstat64(path, buf);
-	if (ret == -1)
-		goto exit;
-
-	uid = next_geteuid();
-	if (buf->st_uid == uid)
-		buf->st_uid = geteuid();
-
-	gid = getegid();
-	if (buf->st_gid == gid)
-		buf->st_gid = 0;
-
-exit:
-	return ret;
-}
-
-__attribute__((visibility("hidden")))
 int rootfstatat64(int fd, const char *path, struct stat64 *buf, int flags)
 {
 	uid_t uid;
@@ -501,75 +281,6 @@ int rootfstatat64(int fd, const char *path, struct stat64 *buf, int flags)
 	int ret;
 
 	ret = next_fstatat64(fd, path, buf, flags);
-	if (ret == -1)
-		goto exit;
-
-	uid = next_geteuid();
-	if (buf->st_uid == uid)
-		buf->st_uid = geteuid();
-
-	gid = getegid();
-	if (buf->st_gid == gid)
-		buf->st_gid = 0;
-
-exit:
-	return ret;
-}
-
-__attribute__((visibility("hidden")))
-int __rootxstat64(int ver, const char *path, struct stat64 *buf)
-{
-	uid_t uid;
-	gid_t gid;
-	int ret;
-
-	ret = next___xstat64(ver, path, buf);
-	if (ret == -1)
-		goto exit;
-
-	uid = next_geteuid();
-	if (buf->st_uid == uid)
-		buf->st_uid = geteuid();
-
-	gid = getegid();
-	if (buf->st_gid == gid)
-		buf->st_gid = 0;
-
-exit:
-	return ret;
-}
-
-__attribute__((visibility("hidden")))
-int __rootfxstat64(int ver, int fd, struct stat64 *buf)
-{
-	uid_t uid;
-	gid_t gid;
-	int ret;
-
-	ret = next___fxstat64(ver, fd, buf);
-	if (ret == -1)
-		goto exit;
-
-	uid = next_geteuid();
-	if (buf->st_uid == uid)
-		buf->st_uid = geteuid();
-
-	gid = getegid();
-	if (buf->st_gid == gid)
-		buf->st_gid = 0;
-
-exit:
-	return ret;
-}
-
-__attribute__((visibility("hidden")))
-int __rootlxstat64(int ver, const char *path, struct stat64 *buf)
-{
-	uid_t uid;
-	gid_t gid;
-	int ret;
-
-	ret = next___lxstat64(ver, path, buf);
 	if (ret == -1)
 		goto exit;