+++ fstat.c
@@ -7,36 +7,14 @@
 #include <stdio.h>
 #include <errno.h>
 #include <limits.h>
+#include <fcntl.h>
 #include <dlfcn.h>
 
 #include <sys/stat.h>
 
 #include "iamroot.h"
 
-extern int rootfstat(int, struct stat *);
-
-__attribute__((visibility("hidden")))
-int next_fstat(int fd, struct stat *statbuf)
-{
-	int (*sym)(int, struct stat *);
-	int ret;
-
-	sym = dlsym(RTLD_NEXT, "fstat");
-	if (!sym) {
-		int next___fxstat(int, int, struct stat *);
-#if defined(__arm__)
-		return next___fxstat(3, fd, statbuf);
-#else
-		return next___fxstat(0, fd, statbuf);
-#endif
-	}
-
-	ret = sym(fd, statbuf);
-	if (ret == -1)
-		__fpathperror(fd, __func__);
-
-	return ret;
-}
+extern int rootfstatat(int, const char *, struct stat *, int);
 
 int fstat(int fd, struct stat *statbuf)
 {
@@ -52,7 +30,7 @@ int fstat(int fd, struct stat *statbuf)
 
 	__debug("%s(fd: %i <-> %s, ...)\n", __func__, fd, buf);
 
-	return rootfstat(fd, statbuf);
+	return rootfstatat(fd, "", statbuf, AT_EMPTY_PATH);
 }
 
 #ifdef __GLIBC__