+++ fstat64.c
@@ -7,6 +7,7 @@
 #include <stdio.h>
 #include <errno.h>
 #include <limits.h>
+#include <fcntl.h>
 #include <dlfcn.h>
 
 #include <sys/stat.h>
@@ -14,30 +15,7 @@
 #include "iamroot.h"
 
 #ifdef __GLIBC__
-extern int rootfstat64(int, struct stat64 *);
-
-__attribute__((visibility("hidden")))
-int next_fstat64(int fd, struct stat64 *statbuf)
-{
-	int (*sym)(int, struct stat64 *);
-	int ret;
-
-	sym = dlsym(RTLD_NEXT, "fstat64");
-	if (!sym) {
-		int next___fxstat64(int, int, struct stat64 *);
-#if defined(__arm__)
-		return next___fxstat64(3, fd, statbuf);
-#else
-		return next___fxstat64(0, fd, statbuf);
-#endif
-	}
-
-	ret = sym(fd, statbuf);
-	if (ret == -1)
-		__fpathperror(fd, __func__);
-
-	return ret;
-}
+extern int rootfstatat64(int, const char *, struct stat64 *, int);
 
 int fstat64(int fd, struct stat64 *statbuf)
 {
@@ -53,7 +31,7 @@ int fstat64(int fd, struct stat64 *statbuf)
 
 	__debug("%s(fd: %i <-> '%s', ...)\n", __func__, fd, buf);
 
-	return rootfstat64(fd, statbuf);
+	return rootfstatat64(fd, "", statbuf, AT_EMPTY_PATH);
 }
 
 weak_alias(fstat64, __fstat64);