+++ lstat64.c
@@ -15,30 +15,7 @@
 #include "iamroot.h"
 
 #ifdef __GLIBC__
-extern int rootlstat64(const char *, struct stat64 *);
-
-__attribute__((visibility("hidden")))
-int next_lstat64(const char *path, struct stat64 *statbuf)
-{
-	int (*sym)(const char *, struct stat64 *);
-	int ret;
-
-	sym = dlsym(RTLD_NEXT, "lstat64");
-	if (!sym) {
-		int next___lxstat64(int, const char *, struct stat64 *);
-#if defined(__arm__)
-		return next___lxstat64(3, path, statbuf);
-#else
-		return next___lxstat64(0, path, statbuf);
-#endif
-	}
-
-	ret = sym(path, statbuf);
-	if (ret == -1)
-		__pathperror(path, __func__);
-
-	return ret;
-}
+extern int rootfstatat64(int, const char *, struct stat64 *, int);
 
 int lstat64(const char *path, struct stat64 *statbuf)
 {
@@ -54,7 +31,7 @@ int lstat64(const char *path, struct stat64 *statbuf)
 
 	__debug("%s(path: '%s' -> '%s', ...)\n", __func__, path, buf);
 
-	return rootlstat64(buf, statbuf);
+	return rootfstatat64(AT_FDCWD, buf, statbuf, AT_SYMLINK_NOFOLLOW);
 }
 
 weak_alias(lstat64, __lstat64);