+++ stat.c
@@ -14,30 +14,7 @@
 
 #include "iamroot.h"
 
-extern int rootstat(const char *, struct stat *);
-
-__attribute__((visibility("hidden")))
-int next_stat(const char *path, struct stat *statbuf)
-{
-	int (*sym)(const char *, struct stat *);
-	int ret;
-
-	sym = dlsym(RTLD_NEXT, "stat");
-	if (!sym) {
-		int next___xstat(int, const char *, struct stat *);
-#if defined(__arm__)
-		return next___xstat(3, path, statbuf);
-#else
-		return next___xstat(0, path, statbuf);
-#endif
-	}
-
-	ret = sym(path, statbuf);
-	if (ret == -1)
-		__pathperror(path, __func__);
-
-	return ret;
-}
+extern int rootfstatat(int, const char *, struct stat *, int);
 
 int stat(const char *path, struct stat *statbuf)
 {
@@ -52,5 +29,5 @@ int stat(const char *path, struct stat *statbuf)
 
 	__debug("%s(path: '%s' -> '%s', ...)\n", __func__, path, buf);
 
-	return rootstat(buf, statbuf);
+	return rootfstatat(AT_FDCWD, buf, statbuf, 0);
 }